requirements.txt
adabelief-pytorch==0.2.1
aiosignal==1.3.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filelock==3.12.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.5.0
grpcio==1.58.0
huggingface-hub==0.14.1
idna==3.4
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.24.3
pure-eval==0.2.2
pyarrow==13.0.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
ray==2.6.3
referencing==0.30.2
regex==2023.5.5
requests==2.31.0
rpds-py==0.10.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
